Refactor 1st draft complete
